# gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: TaskManager API
  title: TaskManager API
  version: 1.0.0
servers:
- url: https://TaskManager
paths:
...
  /tasks/{id}:
    delete:
      operationId: deleteTaskById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: number
        style: simple
      responses:
        "200":
          description: OK
      summary: DELETE tasks/{id}
      x-accepts: application/json
    get:
      operationId: getTaskById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: number
        style: simple
      responses:
        "200":
          description: OK
      summary: GET tasks/{id}
      x-accepts: application/json
...